backend/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# python-jose[cryptography]==3.3.0

# python-multipart==0.0.6
# google-auth==2.23.4
# google-auth-oauthlib==1.1.0
# google-api-python-client==2.108.0
# pymongo==4.6.0
# pyjwt
# cryptography==41.0.7
# python-dotenv==1.0.0
# fastapi-jwt-auth
# argon2-cffi==23.1.0 
# pytesseract==0.3.10
# Pillow==10.0.0
# opencv-python-headless==4.8.1.78
# nltk==3.8.1
# cloudinary==1.37.0
# python-dotenv==1.0.0
# numpy==1.24.3

# fastapi==0.104.1
# uvicorn==0.24.0
# pytesseract==0.3.10
# pillow==10.1.0
# opencv-python-headless==4.8.1.78
# nltk==3.8.1
# python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
httpx==0.25.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
python-multipart==0.0.6
pymongo==4.6.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pyjwt
argon2-cffi==23.1.0
Pillow==10.1.0
nltk==3.8.1
pytesseract==0.3.10
# Skip opencv-python for now - it's heavy